Add functionality to add user to database on registration
