+++ app/src/main/java/edu/gatech/cc/eatsafe/RegistrationActivity.java
@@ -15,7 +15,10 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
 
+import java.util.ArrayList;
 import java.util.regex.Matcher;
 
 import static edu.gatech.cc.eatsafe.LoginActivity.VALID_EMAIL_ADDRESS_REGEX;
@@ -30,6 +33,7 @@ public class RegistrationActivity extends AppCompatActivity {
     private EditText mPasswordView;
 
     private final FirebaseAuth mAuth = FirebaseAuth.getInstance();
+    private final DatabaseReference mDatabase = FirebaseDatabase.getInstance().getReference();
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
@@ -108,6 +112,7 @@ public class RegistrationActivity extends AppCompatActivity {
 
 
         if (!cancel) {
+            // Create user authentication
             mAuth.createUserWithEmailAndPassword(email, password)
                     .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
                         @Override
@@ -131,6 +136,12 @@ public class RegistrationActivity extends AppCompatActivity {
                             }
                         }
                     });
+
+            // Add user to database
+            DatabaseReference reference = mDatabase.child("users").child(email).push();
+            reference.setValue(new UserInformation("", "", null, email,
+                    new ArrayList<String>()));
+
         } else {
             focusView.requestFocus();
         }
